tidy: Use the correct type for loop counters.
The clang-tidy "loop variable too small" check pointed out many places
where the variable used in a for loop is too small for upper limit on
the value being counted.  Almost all of these are for using a uint to
count up to x.size().  Since the size() function returns a value of
type size_t, the trivial solution to these is to change the counter
type to size_t so its always guaranteed to be the right size.

https://clang.llvm.org/extra/clang-tidy/checks/bugprone-too-small-loop-variable.html

+++ mythplugins/mytharchive/mytharchive/recordingselector.cpp
@@ -517,7 +517,7 @@ void RecordingSelector::updateSelectedList()
     for (int x = 0; x < m_archiveList->size(); x++)
     {
         ArchiveItem *a = m_archiveList->at(x);
-        for (uint y = 0; y < m_recordingList->size(); y++)
+        for (size_t y = 0; y < m_recordingList->size(); y++)
         {
             ProgramInfo *p = m_recordingList->at(y);
             if (p->GetPlaybackURL(false, true) == a->filename)
