+++ mythtv/libs/libmythtv/channelscan/channelimporter.cpp
@@ -89,7 +89,7 @@ void ChannelImporter::Process(const ScanDTVTransportList &_transports,
     if (m_do_delete)
     {
         ScanDTVTransportList trans = transports;
-        for (uint i = 0; i < db_trans.size(); ++i)
+        for (size_t i = 0; i < db_trans.size(); ++i)
             trans.push_back(db_trans[i]);
         uint deleted_count = DeleteChannels(trans);
         if (deleted_count)
@@ -146,9 +146,9 @@ uint ChannelImporter::DeleteChannels(
     vector<uint> off_air_list;
     QMap<uint,bool> deleted;
 
-    for (uint i = 0; i < transports.size(); ++i)
+    for (size_t i = 0; i < transports.size(); ++i)
     {
-        for (uint j = 0; j < transports[i].m_channels.size(); ++j)
+        for (size_t j = 0; j < transports[i].m_channels.size(); ++j)
         {
             ChannelInsertInfo chan = transports[i].m_channels[j];
             bool was_in_db = chan.m_db_mplexid && chan.m_channel_id;
@@ -176,7 +176,7 @@ uint ChannelImporter::DeleteChannels(
 
     if (kDeleteAll == action)
     {
-        for (uint k = 0; k < off_air_list.size(); ++k)
+        for (size_t k = 0; k < off_air_list.size(); ++k)
         {
             int i = off_air_list[k] >> 16, j = off_air_list[k] & 0xFFFF;
             ChannelUtil::DeleteChannel(
@@ -186,7 +186,7 @@ uint ChannelImporter::DeleteChannels(
     }
     else if (kDeleteInvisibleAll == action)
     {
-        for (uint k = 0; k < off_air_list.size(); ++k)
+        for (size_t k = 0; k < off_air_list.size(); ++k)
         {
             int i = off_air_list[k] >> 16, j = off_air_list[k] & 0xFFFF;
             int chanid = transports[i].m_channels[j].m_channel_id;
@@ -207,11 +207,11 @@ uint ChannelImporter::DeleteChannels(
         return 0;
 
     // Create a new transports list without the deleted channels
-    for (uint i = 0; i < transports.size(); ++i)
+    for (size_t i = 0; i < transports.size(); ++i)
     {
         newlist.push_back(transports[i]);
         newlist.back().m_channels.clear();
-        for (uint j = 0; j < transports[i].m_channels.size(); ++j)
+        for (size_t j = 0; j < transports[i].m_channels.size(); ++j)
         {
             if (!deleted.contains(i<<16|j))
             {
@@ -383,14 +383,14 @@ ScanDTVTransportList ChannelImporter::InsertChannels(
 
     // insert all channels with non-conflicting channum
     // and complete tuning information.
-    for (uint i = 0; i < transports.size(); ++i)
+    for (size_t i = 0; i < transports.size(); ++i)
     {
         bool created_new_transport = false;
         ScanDTVTransport new_transport;
         bool created_filter_transport = false;
         ScanDTVTransport filter_transport;
 
-        for (uint j = 0; j < transports[i].m_channels.size(); ++j)
+        for (size_t j = 0; j < transports[i].m_channels.size(); ++j)
         {
             ChannelInsertInfo chan = transports[i].m_channels[j];
 
@@ -593,14 +593,14 @@ ScanDTVTransportList ChannelImporter::UpdateChannels(
 
     // update all channels with non-conflicting channum
     // and complete tuning information.
-    for (uint i = 0; i < transports.size(); ++i)
+    for (size_t i = 0; i < transports.size(); ++i)
     {
         bool created_transport = false;
         ScanDTVTransport new_transport;
         bool created_filter_transport = false;
         ScanDTVTransport filter_transport;
 
-        for (uint j = 0; j < transports[i].m_channels.size(); ++j)
+        for (size_t j = 0; j < transports[i].m_channels.size(); ++j)
         {
             ChannelInsertInfo chan = transports[i].m_channels[j];
 
@@ -751,12 +751,12 @@ void ChannelImporter::CleanupDuplicates(ScanDTVTransportList &transports) const
 
     vector<bool> ignore;
     ignore.resize(transports.size());
-    for (uint i = 0; i < transports.size(); ++i)
+    for (size_t i = 0; i < transports.size(); ++i)
     {
         if (ignore[i])
             continue;
 
-        for (uint j = i+1; j < transports.size(); ++j)
+        for (size_t j = i+1; j < transports.size(); ++j)
         {
             if (!transports[i].IsEqual(
                     tuner_type, transports[j], 500 * freq_mult))
@@ -764,10 +764,10 @@ void ChannelImporter::CleanupDuplicates(ScanDTVTransportList &transports) const
                 continue;
             }
 
-            for (uint k = 0; k < transports[j].m_channels.size(); ++k)
+            for (size_t k = 0; k < transports[j].m_channels.size(); ++k)
             {
                 bool found_same = false;
-                for (uint l = 0; l < transports[i].m_channels.size(); ++l)
+                for (size_t l = 0; l < transports[i].m_channels.size(); ++l)
                 {
                     if (transports[j].m_channels[k].IsSameChannel(
                             transports[i].m_channels[l]))
@@ -793,10 +793,10 @@ void ChannelImporter::FilterServices(ScanDTVTransportList &transports) const
     bool require_av = (m_service_requirements & kRequireAV) == kRequireAV;
     bool require_a  = m_service_requirements & kRequireAudio;
 
-    for (uint i = 0; i < transports.size(); ++i)
+    for (size_t i = 0; i < transports.size(); ++i)
     {
         ChannelInsertInfoList filtered;
-        for (uint k = 0; k < transports[i].m_channels.size(); ++k)
+        for (size_t k = 0; k < transports[i].m_channels.size(); ++k)
         {
             if (m_fta_only && transports[i].m_channels[k].m_is_encrypted &&
                 transports[i].m_channels[k].m_decryption_status != kEncDecrypted)
@@ -868,17 +868,17 @@ ScanDTVTransportList ChannelImporter::GetDBTransports(
         bool newt_found = false;
         QMap<uint,bool> found_chan;
 
-        for (uint i = 0; i < transports.size(); ++i)
+        for (size_t i = 0; i < transports.size(); ++i)
         {
             if (!transports[i].IsEqual(tuner_type, newt, 500 * freq_mult, true))
                 continue;
 
             transports[i].m_mplex = mplexid;
             newt_found = true;
-            for (uint j = 0; j < transports[i].m_channels.size(); ++j)
+            for (size_t j = 0; j < transports[i].m_channels.size(); ++j)
             {
                 ChannelInsertInfo &chan = transports[i].m_channels[j];
-                for (uint k = 0; k < newt.m_channels.size(); ++k)
+                for (size_t k = 0; k < newt.m_channels.size(); ++k)
                 {
                     if (newt.m_channels[k].IsSameChannel(chan, true))
                     {
@@ -905,7 +905,7 @@ ScanDTVTransportList ChannelImporter::GetDBTransports(
             ScanDTVTransport tmp = newt;
             tmp.m_channels.clear();
 
-            for (uint k = 0; k < newt.m_channels.size(); ++k)
+            for (size_t k = 0; k < newt.m_channels.size(); ++k)
             {
                 if (!found_chan[k])
                     tmp.m_channels.push_back(newt.m_channels[k]);
@@ -922,9 +922,9 @@ ScanDTVTransportList ChannelImporter::GetDBTransports(
 void ChannelImporter::FixUpOpenCable(ScanDTVTransportList &transports)
 {
     ChannelImporterBasicStats info;
-    for (uint i = 0; i < transports.size(); ++i)
+    for (size_t i = 0; i < transports.size(); ++i)
     {
-        for (uint j = 0; j < transports[i].m_channels.size(); ++j)
+        for (size_t j = 0; j < transports[i].m_channels.size(); ++j)
         {
             ChannelInsertInfo &chan = transports[i].m_channels[j];
             if (((chan.m_could_be_opencable && (chan.m_si_standard == "mpeg")) ||
@@ -940,9 +940,9 @@ ChannelImporterBasicStats ChannelImporter::CollectStats(
     const ScanDTVTransportList &transports)
 {
     ChannelImporterBasicStats info;
-    for (uint i = 0; i < transports.size(); ++i)
+    for (size_t i = 0; i < transports.size(); ++i)
     {
-        for (uint j = 0; j < transports[i].m_channels.size(); ++j)
+        for (size_t j = 0; j < transports[i].m_channels.size(); ++j)
         {
             const ChannelInsertInfo &chan = transports[i].m_channels[j];
             int enc = (chan.m_is_encrypted) ?
@@ -981,9 +981,9 @@ ChannelImporterUniquenessStats ChannelImporter::CollectUniquenessStats(
 {
     ChannelImporterUniquenessStats stats;
 
-    for (uint i = 0; i < transports.size(); ++i)
+    for (size_t i = 0; i < transports.size(); ++i)
     {
-        for (uint j = 0; j < transports[i].m_channels.size(); ++j)
+        for (size_t j = 0; j < transports[i].m_channels.size(); ++j)
         {
             const ChannelInsertInfo &chan = transports[i].m_channels[j];
             stats.m_unique_prognum +=
@@ -1144,8 +1144,8 @@ QString ChannelImporter::FormatChannels(
 {
     QString msg;
 
-    for (uint i = 0; i < transports.size(); ++i)
-        for (uint j = 0; j < transports[i].m_channels.size(); ++j)
+    for (size_t i = 0; i < transports.size(); ++i)
+        for (size_t j = 0; j < transports[i].m_channels.size(); ++j)
             msg += FormatChannel(transports[i], transports[i].m_channels[j],
                                  &info) + "\n";
 
@@ -1243,9 +1243,9 @@ void ChannelImporter::CountChannels(
     ChannelType type, uint &new_chan, uint &old_chan)
 {
     new_chan = old_chan = 0;
-    for (uint i = 0; i < transports.size(); ++i)
+    for (size_t i = 0; i < transports.size(); ++i)
     {
-        for (uint j = 0; j < transports[i].m_channels.size(); ++j)
+        for (size_t j = 0; j < transports[i].m_channels.size(); ++j)
         {
             ChannelInsertInfo chan = transports[i].m_channels[j];
             if (IsType(info, chan, type))
