+++ mythtv/libs/libmythtv/channelsettings.cpp
@@ -350,7 +350,7 @@ class CommMethod : public MythUIComboBoxSetting
         tmp.push_front(COMM_DETECT_UNINIT);
         tmp.push_back(COMM_DETECT_COMMFREE);
 
-        for (uint i = 0; i < tmp.size(); i++)
+        for (size_t i = 0; i < tmp.size(); i++)
             addSelection(SkipTypeToString(tmp[i]), QString::number(tmp[i]));
     }
 };
